grammar edits
old->
Each year, the U.S. sees approximately $1billion in property and crop
damage.
new ->
Each year, the U.S. sees approximately $1 billion in property and crop
damage.

old ->
The practical applications are many, ranging from mitigating risk of
encountering adverse storm events (e.g. when not to go to the beach) to
anticipating the best time start growing certain crops.
new ->
The practical applications are many, ranging from mitigating risk of
encountering adverse storm events (e.g. when not to go to the beach) to
anticipating the best time to start growing certain crops.

old ->
Then based on the hail size, we’ll produce dummy variables are produced
for each 2+ inch and 3+ inch thresholds.
new ->
Then based on the hail size, we’ll produce dummy variables to be
produced for each 2+ inch and 3+ inch thresholds.

old ->
Note that climatologies are calculating based on day-grid cells (e.g.
whether a hail signature was detected in a 0.25 degree grid cell in a
given day) as opposed to the raw hail signatures.
new ->
Note that climatologies are calculated based on day-grid cells (e.g.
whether a hail signature was detected in a 0.25 degree grid cell in a
given day) as opposed to the raw hail signatures.

old ->
To do this for grid points, we’ll spatially join the grid points to the
closest county, that is see which county a grid point falls.
new ->
To do this for grid points, we’ll spatially join the grid points to the
closest county, that is to see which county a grid point falls in.